% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_dissimilar.R
\name{check_dissimilar}
\alias{check_dissimilar}
\title{Check for dissimilar rows}
\usage{
check_dissimilar(x, tool, threshold = 7, ...)
}
\arguments{
\item{x}{\emph{A pointblank agent or a data table}

\verb{obj:<ptblank_agent>|obj:<tbl_*>} // \strong{required}

A data frame, tibble (\code{tbl_df} or \code{tbl_dbi}), Spark DataFrame
(\code{tbl_spark}), or, an \emph{agent} object of class \code{ptblank_agent} that is
commonly created with \code{\link[pointblank:create_agent]{create_agent()}}.}

\item{tool}{kobo survey}

\item{threshold}{minimum number of dissimilar columns between rows}

\item{...}{Unused for now, kept for future extensions}
}
\value{
For the validation function, the return value is either a
\code{ptblank_agent} object or a table object (depending on whether an agent
object or a table was passed to \code{x}). The expectation function invisibly
returns its input but, in the context of testing data, the function is
called primarily for its potential side-effects (e.g., signaling failure).
The test function returns a logical value.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
