% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui.R
\name{download_submissions_mock}
\alias{download_submissions_mock}
\title{Mock function to simulate data submission download from robotoolbox}
\usage{
download_submissions_mock(uid)
}
\arguments{
\item{uid}{A character string representing the unique identifier.}
}
\value{
A data frame if \code{uid} is "valid_uid", otherwise throws an error.
}
\description{
This function simulates a network delay and returns a dummy data frame
if a specific UID is provided. Otherwise, it throws an error.
}
\keyword{internal}
