% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_uuid.R
\name{identify_uuid_column}
\alias{identify_uuid_column}
\title{Identify a UUID column}
\usage{
identify_uuid_column(x)
}
\arguments{
\item{x}{\emph{A pointblank agent or a data table}

\verb{obj:<ptblank_agent>|obj:<tbl_*>} // \strong{required}

A data frame, tibble (\code{tbl_df} or \code{tbl_dbi}), Spark DataFrame
(\code{tbl_spark}), or, an \emph{agent} object of class \code{ptblank_agent} that is
commonly created with \code{\link[pointblank:create_agent]{create_agent()}}.}
}
\value{
The name of the first UUID column found, or \code{NA_character_} if none
are found.
}
\description{
This function looks for a UUID column in a data frame by
checking for common names that UUID columns might have, such as \verb{_uuid},
\code{uuid}, or \code{X_uuid}.
}
\details{
The function uses the \code{fmatch} function to find the first match
among the potential UUID column names. If no match is found, it returns
\code{NA_character_}.
}
\examples{
df <- data.frame(
  `_uuid` = c(
    "123e4567-e89b-12d3-a456-426614174000",
    "123e4567-e89b-12d3-a456-426614174001"
  ),
  value = c(1, 2)
)

identify_uuid_column(df)
}
