% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KQC-package.R
\docType{package}
\name{KQC-package}
\alias{KQC}
\alias{KQC-package}
\title{KQC: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://pkgs.impact-initiatives.org/KQC}
}

}
\author{
\strong{Maintainer}: Iyed GHEDAMSI \email{iyed.ghedamsi@impact-initiatives.org}

Other contributors:
\itemize{
  \item IMPACT Initiatives [contributor]
}

}
\keyword{internal}
