% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui.R
\name{kqc_module_server}
\alias{kqc_module_server}
\title{Server Logic for Data Quality Control Module}
\usage{
kqc_module_server(id)
}
\arguments{
\item{id}{A unique character string identifying the module instance.}
}
\value{
A reactive expression returning the validated data frame.
}
\description{
Handles downloading data and validating it using pointblank.
}
