% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui.R
\name{kqc_module_ui}
\alias{kqc_module_ui}
\title{User Interface for Data Quality Control Module}
\usage{
kqc_module_ui(id)
}
\arguments{
\item{id}{A unique character string identifying the module instance.}
}
\value{
A Shiny tag list for inclusion in a UI.
}
\description{
User Interface for Data Quality Control Module
}
