% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui.R
\name{kqc_ui}
\alias{kqc_ui}
\title{User Interface for Data Submission Validator}
\usage{
kqc_ui()
}
\value{
A Shiny miniUI application for validating data submissions.
}
\description{
This function launches a Shiny miniUI application for validating data
submissions.
}
