% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{msna_opt_cl_HH}
\alias{msna_opt_cl_HH}
\title{Cleaning log for household data from the 2021 oPt MSNA}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 1273 rows and 18 columns.
}
\source{
See \code{\link{msna_opt_HH}}.
}
\usage{
msna_opt_cl_HH
}
\description{
This dataset records all manual or scripted corrections applied to the household dataset. It follows a structured cleaning log format.
}
\keyword{datasets}
