% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{msna_opt_dl_HH}
\alias{msna_opt_dl_HH}
\title{Deletion log for household records in the 2021 oPt MSNA}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 46 rows and 10 columns.
}
\source{
See \code{\link{msna_opt_HH}}.
}
\usage{
msna_opt_dl_HH
}
\description{
This log identifies records removed from the dataset during data cleaning, typically due to quality issues or duplication.
}
\keyword{datasets}
