% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{msna_opt_HH}
\alias{msna_opt_HH}
\title{Household-level data from the 2021 oPt MSNA}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 7514 rows and 847 columns.
}
\source{
\href{https://repository.impact-initiatives.org/document/impact/efc7b590/REACH_oPt_MSNA-Dataset_OPT2101_16082021.xlsx}{IMPACT Initiatives Resource Centre}
}
\usage{
msna_opt_HH
}
\description{
This dataset contains household-level survey data from the 2021 Multi-Sector Needs Assessment (MSNA)
conducted in the occupied Palestinian territory (oPt). It includes socioeconomic indicators, access to services,
perceived needs, coping mechanisms, and household demographics.
}
\keyword{datasets}
