% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{msna_opt_Ind}
\alias{msna_opt_Ind}
\title{Individual-level data from the 2021 oPt MSNA}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 2678 rows and 9 columns.
}
\source{
See \code{\link{msna_opt_HH}}.
}
\usage{
msna_opt_Ind
}
\description{
This dataset includes information on individuals within surveyed households. It allows disaggregation by age, gender,
and other demographic characteristics.
}
\keyword{datasets}
