% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui.R
\name{kqc_module_server}
\alias{kqc_module_server}
\title{Server logic for the Data Quality Control Module}
\usage{
kqc_module_server(id)
}
\arguments{
\item{id}{character. A unique ID for the module instance.}
}
\value{
A reactive expression that returns the pointblank agent object.
}
\description{
Server logic for the Data Quality Control Module
}
