% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui.R
\name{kqc_module_ui}
\alias{kqc_module_ui}
\title{UI for the Data Quality Control Module}
\usage{
kqc_module_ui(id)
}
\arguments{
\item{id}{character. A unique ID for the module instance.}
}
\value{
A Shiny UI tag list.
}
\description{
UI for the Data Quality Control Module
}
